# openapi_temp.json
{"openapi":"3.1.0","info":{"title":"Infinity Tag Backend","description":"Infinity Tagï¼æ æ­¢ä¾¿ç­¾ï¼Web åç«¯ API","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["ç³»ç»"],"summary":"Health Check","description":"å¥åº·æ£æ¥æ¥å£","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["ç³»ç»"],"summary":"Root","description":"æ ¹è·¯å¾","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/activate":{"post":{"tags":["Auth"],"summary":"Activate Device","description":"è®¾å¤æ¿æ´»/æ³¨åæ¥å£\nè¿å: Token + Device Secret (ä»ä¸æ¬¡)","operationId":"activate_device_api_v1_auth_activate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceActivateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"è®¾å¤ç»å½æ¥å£\nè¿å: ä» Token","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Auth"],"summary":"Read Users Me","description":"è·åå½åç¨æ·ä¿¡æ¯","operationId":"read_users_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/profile/":{"get":{"tags":["Profile"],"summary":"Get Profile","description":"è·åå½åç¨æ·çæ¡£æ¡ä¿¡æ¯","operationId":"get_profile_api_v1_profile__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Profile"],"summary":"Create Or Update Profile","description":"åå»ºæå®å¨æ´æ°ç¨æ·æ¡£æ¡","operationId":"create_or_update_profile_api_v1_profile__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"patch":{"tags":["Profile"],"summary":"Update Profile Partial","description":"é¨åæ´æ°ç¨æ·æ¡£æ¡","operationId":"update_profile_partial_api_v1_profile__patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/almanac/generate":{"post":{"tags":["Almanac"],"summary":"Generate Almanac","description":"çæé»å (é»è®¤ä»æ¥)","operationId":"generate_almanac_api_v1_almanac_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlmanacGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlmanacResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/almanac/history":{"get":{"tags":["Almanac"],"summary":"Get History","description":"è·ååå²é»å","operationId":"get_history_api_v1_almanac_history_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":30,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlmanacHistoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/divination/words":{"get":{"tags":["Divination"],"summary":"Get Random Words","description":"è·åéæºå¤éå­","operationId":"get_random_words_api_v1_divination_words_get","parameters":[{"name":"count","in":"query","required":false,"schema":{"type":"integer","default":8,"title":"Count"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Response Get Random Words Api V1 Divination Words Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/divination/interpret":{"post":{"tags":["Divination"],"summary":"Interpret Divination","description":"è§£å­æµç®","operationId":"interpret_divination_api_v1_divination_interpret_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DivinationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DivinationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/renderer/preview":{"get":{"tags":["Renderer"],"summary":"Preview Almanac","description":"é¢è§ä»æ¥é»åæ¸²æææ (PNG)","operationId":"preview_almanac_api_v1_renderer_preview_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"target_date","in":"query","required":false,"schema":{"type":"string","format":"date","default":"2026-02-04","title":"Target Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/renderer/bitmap":{"get":{"tags":["Renderer"],"summary":"Get Almanac Bitmap","description":"è·å ESP32 ç¨çåå§ä½å¾æ°æ® (Binary)","operationId":"get_almanac_bitmap_api_v1_renderer_bitmap_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"target_date","in":"query","required":false,"schema":{"type":"string","format":"date","default":"2026-02-04","title":"Target Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/images/preview":{"post":{"tags":["Images"],"summary":"Preview Image","description":"é¢è§å¾çå¤çææ (ä¸ä¿å­)\nåç«¯å¯ä»¥åéåå¾ + åæ°ï¼å®æ¶æ¥çææ","operationId":"preview_image_api_v1_images_preview_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_preview_image_api_v1_images_preview_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/images/":{"get":{"tags":["Images"],"summary":"Get Images","description":"è·åç¨æ·çææèªå®ä¹å¾ç","operationId":"get_images_api_v1_images__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ImageOut"},"type":"array","title":"Response Get Images Api V1 Images  Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Images"],"summary":"Upload Image","description":"ä¸ä¼ å¾ç (æ¯æè£åªãäºå¼åç­åæ°)","operationId":"upload_image_api_v1_images__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_image_api_v1_images__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/images/{image_id}":{"delete":{"tags":["Images"],"summary":"Delete Image","description":"å é¤å¾ç","operationId":"delete_image_api_v1_images__image_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"integer","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/images/reorder":{"put":{"tags":["Images"],"summary":"Reorder Images","description":"éæ°æåº","operationId":"reorder_images_api_v1_images_reorder_put","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ImageReorder"},"type":"array","title":"Orders"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/images/{image_id}/bitmap":{"get":{"tags":["Images"],"summary":"Get Image Bitmap","description":"è·åå¾ççåå§ä½å¾æ°æ® (ä¾ ESP32 ä¸è½½)","operationId":"get_image_bitmap_api_v1_images__image_id__bitmap_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"integer","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActivateResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"device_secret":{"type":"string","title":"Device Secret","description":"è®¾å¤HMACå¯é¥ (ä»é¦æ¬¡è¿åï¼è¯·å¦¥åä¿å­)"},"user_id":{"type":"integer","title":"User Id"},"device_id":{"type":"string","title":"Device Id"}},"type":"object","required":["access_token","device_secret","user_id","device_id"],"title":"ActivateResponse","description":"æ¿æ´»ååº - è¿å Token å è®¾å¤å¯é¥"},"AlmanacGenerateRequest":{"properties":{"date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date","description":"è¯·æ±æ¥æï¼é»è®¤ä»æ¥"}},"type":"object","title":"AlmanacGenerateRequest"},"AlmanacHistoryResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/AlmanacResponse"},"type":"array","title":"Data"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["data","total"],"title":"AlmanacHistoryResponse"},"AlmanacResponse":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"lunar_date":{"type":"string","title":"Lunar Date"},"ganzhi_year":{"type":"string","title":"Ganzhi Year"},"ganzhi_month":{"type":"string","title":"Ganzhi Month"},"ganzhi_day":{"type":"string","title":"Ganzhi Day"},"favorable":{"items":{"type":"string"},"type":"array","title":"Favorable"},"unfavorable":{"items":{"type":"string"},"type":"array","title":"Unfavorable"},"lucky_direction":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lucky Direction"},"lucky_item":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lucky Item"},"energy_level":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Energy Level"},"commentary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Commentary"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["date","lunar_date","ganzhi_year","ganzhi_month","ganzhi_day","favorable","unfavorable","lucky_direction","lucky_item","energy_level","commentary","generated_at"],"title":"AlmanacResponse"},"Body_preview_image_api_v1_images_preview_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"options":{"type":"string","title":"Options","default":"{}"}},"type":"object","required":["file"],"title":"Body_preview_image_api_v1_images_preview_post"},"Body_upload_image_api_v1_images__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"options":{"type":"string","title":"Options","default":"{}"}},"type":"object","required":["file"],"title":"Body_upload_image_api_v1_images__post"},"DeviceActivateRequest":{"properties":{"device_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Device Id","description":"è®¾å¤ID (6ä½ç­ç )ï¼ä¸ºç©ºåèªå¨çæ"},"password":{"type":"string","minLength":6,"title":"Password","description":"ç¨æ·è®¾ç½®çå¯ç "}},"type":"object","required":["password"],"title":"DeviceActivateRequest"},"DivinationRequest":{"properties":{"mode":{"type":"string","pattern":"^(æ¬å½|å®¢åº§)$","title":"Mode","description":"æµç®æ¨¡å¼"},"intent":{"type":"string","maxLength":10,"title":"Intent","description":"æ±æµæå¾ï¼å¦'äºä¸'"},"words":{"items":{"type":"string"},"type":"array","maxItems":2,"minItems":2,"title":"Words","description":"éä¸­çä¸¤ä¸ªå­"}},"type":"object","required":["mode","intent","words"],"title":"DivinationRequest"},"DivinationResponse":{"properties":{"id":{"type":"integer","title":"Id"},"mode":{"type":"string","title":"Mode"},"intent":{"type":"string","title":"Intent"},"selected_words":{"items":{"type":"string"},"type":"array","title":"Selected Words"},"result_idiom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result Idiom"},"result_interpretation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result Interpretation"},"result_advice":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result Advice"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","mode","intent","selected_words","result_idiom","result_interpretation","result_advice","created_at"],"title":"DivinationResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageOut":{"properties":{"display_order":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Display Order","default":0},"id":{"type":"integer","title":"Id"},"url":{"type":"string","title":"Url"},"view_count":{"type":"integer","title":"View Count"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","url","view_count","created_at"],"title":"ImageOut"},"ImageReorder":{"properties":{"id":{"type":"integer","title":"Id"},"new_order":{"type":"integer","title":"New Order"}},"type":"object","required":["id","new_order"],"title":"ImageReorder"},"LoginRequest":{"properties":{"device_id":{"type":"string","title":"Device Id","description":"è®¾å¤ID"},"password":{"type":"string","title":"Password","description":"å¯ç "}},"type":"object","required":["device_id","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"LoginResponse","description":"ç»å½ååº - ä»è¿å Token"},"ProfileCreate":{"properties":{"birth_datetime":{"type":"string","format":"date-time","title":"Birth Datetime"},"birth_place":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Birth Place"},"profession":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profession"},"focus_areas":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Focus Areas"}},"type":"object","required":["birth_datetime"],"title":"ProfileCreate"},"ProfileResponse":{"properties":{"birth_datetime":{"type":"string","format":"date-time","title":"Birth Datetime"},"birth_place":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Birth Place"},"profession":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profession"},"focus_areas":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Focus Areas"},"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"bazi_year":{"type":"string","title":"Bazi Year"},"bazi_month":{"type":"string","title":"Bazi Month"},"bazi_day":{"type":"string","title":"Bazi Day"},"bazi_hour":{"type":"string","title":"Bazi Hour"}},"type":"object","required":["birth_datetime","id","user_id","bazi_year","bazi_month","bazi_day","bazi_hour"],"title":"ProfileResponse"},"ProfileUpdate":{"properties":{"birth_datetime":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Birth Datetime"},"birth_place":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Birth Place"},"profession":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profession"},"focus_areas":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Focus Areas"}},"type":"object","title":"ProfileUpdate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"device_id":{"type":"string","title":"Device Id"},"activated_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Activated At"},"last_login_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Login At"}},"type":"object","required":["id","device_id","activated_at","last_login_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}